simplify printnanny-gst command structure
