+++ gst/src/options.rs
@@ -1,6 +1,46 @@
 use clap::{ArgEnum, PossibleValue};
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ArgEnum)]
+pub enum AppModeOption {
+    // broadcast source video stream over 1 rtp port (light compute)
+    RtpVideo,
+    // broadcast source video, model inference video, and model inference tensor over 3 rtp ports (medium compute)
+    RtpTfliteOverlay,
+    // broadcast video composited from source / inference (heavy compute)
+    RtpTfliteComposite,
+}
+
+impl AppModeOption {
+    pub fn possible_values() -> impl Iterator<Item = PossibleValue<'static>> {
+        AppModeOption::value_variants()
+            .iter()
+            .filter_map(ArgEnum::to_possible_value)
+    }
+}
+
+impl std::fmt::Display for AppModeOption {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        self.to_possible_value()
+            .expect("no values are skipped")
+            .get_name()
+            .fmt(f)
+    }
+}
+
+impl std::str::FromStr for AppModeOption {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        for variant in Self::value_variants() {
+            if variant.to_possible_value().unwrap().matches(s, false) {
+                return Ok(*variant);
+            }
+        }
+        Err(format!("Invalid variant: {}", s))
+    }
+}
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ArgEnum)]
 pub enum SinkOption {
     Fakesink,
     Udpsink,
